requirements.txt
Flask==3.0.2
pytest== 8.1.1 
coverage == 7.4.4
Flask-SQLAlchemy == 3.1.1 
Flask-Migrate == 4.0.7
psycopg2 == 2.9.9